Create scraper.py and model.py
These aren't named particularly well, but the model.py contains some
classes that will be useful in keeping track of what exactly is
scraped. scraper.py has functions and classes designed particularly
for extracting information about a route, specifically how
difficult different climbers found it to be, and which climbers have
completed it.